add `ui/seekable-progress-bar` tests
